merging master into pipeline

+++ backend/src/main/java/com/spring/techpractica/service/session/SessionService.java
@@ -1,19 +1,18 @@
 package com.spring.techpractica.service.session;
 
+import com.spring.techpractica.dto.UserRequestSession;
 import com.spring.techpractica.dto.session.SessionRequest;
+import com.spring.techpractica.dto.session.SessionRequestCreation;
 import com.spring.techpractica.dto.session.SessionResponse;
 import com.spring.techpractica.dto.session.SessionsResponse;
 import com.spring.techpractica.exception.AuthenticationException;
 import com.spring.techpractica.factory.PageRequestFactory;
-import com.spring.techpractica.factory.RequirementFactory;
+import com.spring.techpractica.factory.RequestBuilding;
 import com.spring.techpractica.maper.SessionMapper;
 import com.spring.techpractica.mengmentData.*;
 import com.spring.techpractica.model.SessionRole;
-import com.spring.techpractica.model.entity.AuthenticatedUserSession;
-import com.spring.techpractica.model.entity.Requirement;
-import com.spring.techpractica.model.entity.Session;
-import com.spring.techpractica.model.entity.User;
-import com.spring.techpractica.model.entity.techSkills.Category;
+import com.spring.techpractica.model.entity.*;
+import com.spring.techpractica.model.entity.techSkills.System;
 import com.spring.techpractica.service.session.createSession.CreateSessionService;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -27,19 +26,22 @@ import java.util.stream.Collectors;
 @AllArgsConstructor
 public class SessionService {
 
+    private final RequirementManagementData requirementManagementData;
+
     private final SessionManagementData sessionManagementData;
 
     private final UserManagementData userManagementData;
 
-    private final CategoryManagementData categoryManagementData;
+    private final SystemManagementData systemManagementData;
 
     private final CreateSessionService createSessionService;
 
     private final TechnologyManagementData technologyManagementData;
 
-    private final FieldManagementData fieldManagementData;
+    private final CategoryManagementData categoryManagementData;
 
     private final AuthenticatedUserSessionManagementData authenticatedUserSessionManagementData;
+
 
     public SessionResponse createSession(SessionRequest sessionRequest,
                                          String userEmail) {
@@ -52,7 +54,6 @@ public class SessionService {
     public SessionsResponse getSessionsByUserEmail(String userEmail,
                                                    int pageSize,
                                                    int pageNumber) {
-
 
         User user = userManagementData.getUserByEmail(userEmail);
 
@@ -68,18 +69,19 @@ public class SessionService {
         return null;
     }
 
-    public SessionsResponse getSessionsByCategoryName(String categoryName, int pageSize, int pageNumber) {
+    public SessionsResponse getSessionsBySystemName(String systemName, int pageSize, int pageNumber) {
 
-        Category category = categoryManagementData.getCategoryByName(categoryName);
+        System system = systemManagementData.getSystemByName(systemName);
 
         List<Session> sessions = sessionManagementData
-                .getSessionsByCategoryAndPageable(category,
+                .getSessionsBySystemAndPageable(system,
                         PageRequestFactory.createPageRequest(pageSize, pageNumber));
 
-        long totalSession = sessionManagementData.getNumberOfCategorySessions(category);
+        long totalSession = sessionManagementData.getNumberOfSystemSessions(system);
+
+
 
         return SessionMapper.sessionsAndTotalSessionsToSessionsResponses(sessions, totalSession);
-
     }
 
     public SessionsResponse getUserSessions(String userEmail, int pageSize, int pageNumber) {
@@ -99,8 +101,13 @@ public class SessionService {
             , Long sessionId) {
 
         Session session = sessionManagementData.getSessionById(sessionId);
-        sessionManagementData.deleteSession(session);
+        User user = userManagementData.getUserByEmail(username);
 
+        if (!getUserRole(user.getUserId(), sessionId).equals(SessionRole.OWNER)) {
+            throw new AuthenticationException("User dont have authorization");
+        }
+
+        sessionManagementData.deleteSession(session);
     }
 
     @Transactional
@@ -113,9 +120,11 @@ public class SessionService {
 
         Session session = sessionManagementData.getSessionById(sessionId);
 
-        if (getSessionRole(user.getUserId(), sessionId) != SessionRole.OWNER) {
-            throw new AuthenticationException("User must be an OWNER to perform this action.");
+
+        if (!getUserRole(user.getUserId(), sessionId).equals(SessionRole.OWNER)) {
+            throw new AuthenticationException("User dont have authorization");
         }
+
 
         session.setSessionName(updatedSessionRequest.getNameSession());
 
@@ -128,28 +137,29 @@ public class SessionService {
                         .getTechnologiesByTechnologiesName(updatedSessionRequest.getTechnologies()))
         );
 
-        session.setSessionCategories(
+        session.setSessionSystems(
                 new ArrayList<>(categoriesStringToCategoriesList(
-                        List.of(updatedSessionRequest.getCategory())
+                        List.of(updatedSessionRequest.getSystem())
                 ))
         );
 
-        session.getSessionRequirements().clear();
+//        session.getSessionRequirements().clear();
+//
+//        List<Requirement> requirements = updatedSessionRequest
+//                .getCategories()
+//                .stream()
+//                .map(category -> RequirementFactory.createRequirement(
+//                        session,
+//                        categoryManagementData.getCategoryByCategoryName(category)
+//                ))
+//                .collect(Collectors.toCollection(ArrayList::new));
 
-        List<Requirement> requirements = updatedSessionRequest
-                .getFields()
-                .stream()
-                .map(field -> RequirementFactory.createRequirement(
-                        session,
-                        fieldManagementData.getFieldByFieldName(field)
-                ))
-                .collect(Collectors.toCollection(ArrayList::new));
+//        session.getSessionRequirements()
+//                .addAll(requirements);
 
-        session.getSessionRequirements().addAll(requirements);
-
-        session.setSessionFields(
-                new ArrayList<>(fieldManagementData
-                        .getFieldsByFieldsName(updatedSessionRequest.getFields()))
+        session.setSessionCategories(
+                new ArrayList<>(categoryManagementData
+                        .getCategoriesByCategoriesName(updatedSessionRequest.getCategories()))
         );
 
         sessionManagementData.saveSession(session);
@@ -157,17 +167,17 @@ public class SessionService {
         return SessionMapper.sessionToSessionResponse(session);
     }
 
-    private List<Category> categoriesStringToCategoriesList(List<String> categories) {
+    private List<System> categoriesStringToCategoriesList(List<String> categories) {
         return categories.stream()
-                .map(categoryManagementData::getCategoryByName)
+                .map(systemManagementData::getSystemByName)
                 .toList();
     }
 
-    public SessionRole getSessionRole(Long userId, Long sessionId) {
+    public SessionRole getUserRole(Long userId, Long sessionId) {
 
         AuthenticatedUserSession authenticatedUserSession = authenticatedUserSessionManagementData
                 .findByUserUserIdAndUserSessionId(userId, sessionId)
-                .orElseThrow(() -> new AuthenticationException("User is not authenticated"));
+                .orElseThrow(() -> new AuthenticationException("User Don't have access in session"));
 
         return authenticatedUserSession.getScopedRole();
 
@@ -176,6 +186,64 @@ public class SessionService {
     public SessionsResponse getSessions(int pageSize, int pageNumber) {
         List<Session> sessions = sessionManagementData.
                 getSessionsByPageable(PageRequestFactory.createPageRequest(pageSize, pageNumber));
+
         return SessionMapper.sessionsAndTotalSessionsToSessionsResponses(sessions, sessionManagementData.getNumberOfSessions());
+    }
+
+    @Transactional
+    public void createRequestSession(SessionRequestCreation sessionRequestCreation,
+                                     String userEmail) {
+
+        Session session =
+                sessionManagementData.getSessionById(sessionRequestCreation.getSessionId());
+
+        User user = userManagementData.getUserByEmail(userEmail);
+
+
+        Request request = RequestBuilding.
+                createRequestFrom(session, user, sessionRequestCreation.getBrief());
+
+
+        Requirement requirement =
+                requirementManagementData.
+                        getRequirementBySessionIdAndSystem(session.getSessionId(), sessionRequestCreation.getCategoryName());
+
+        request.setRequirement(requirement);
+
+        session.getSessionRequests()
+                .add(request);
+
+        sessionManagementData.saveSession(session);
+    }
+
+    @Transactional
+    public List<UserRequestSession> getSessionsRequest(Long sessionId, String username) {
+
+        Session session = sessionManagementData.getSessionById(sessionId);
+        User user = userManagementData.getUserByEmail(username);
+
+        if (!getUserRole(user.getUserId(), sessionId).equals(SessionRole.OWNER)) {
+            throw new AuthenticationException("User dont have authorization");
+        }
+
+
+        List<Request> requests = session.getSessionRequests();
+
+
+        if (requests == null) {
+            return new ArrayList<>();
+        }
+
+        return requests
+                .stream()
+                .filter((r) -> r.getRequirement() != null)
+                .map(r -> UserRequestSession.builder()
+                        .brief(r.getBrief())
+                        .username(r.getUser().getUserName())
+                        .categoryName(r.getRequirement()
+                                .getCategory()
+                                .getCategoryName())
+                        .build())
+                .collect(Collectors.toList());
     }
 }
